+++ packages/commonwealth/client/scripts/views/components/PrivyTest/useValueRef.tsx
@@ -1,0 +1,7 @@
+import { useRef } from 'react';
+
+export function useValueRef<Value>(value: Value) {
+  const ref = useRef(value);
+  ref.current = value;
+  return ref;
+}
